update e2fsprogs: add new files
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@63477 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/e2fsprogs/e2fsck/logfile.c
@@ -183,12 +183,12 @@ static void expand_logfn(e2fsck_t ctx, const char *log_fn, struct string *s)
 	}
 }
 
-static int	outbufsize;
+static size_t	outbufsize;
 static void	*outbuf;
 
 static int do_read(int fd)
 {
-	int	c;
+	ssize_t		c;
 	char		*n;
 	char	buffer[4096];
